Load contexts are finally working.
Lots of debugging went into this, and I don't think I'm compeltely
finished yet. Currently, however, all the correct assemblies are
resolved in their respective environments. Next up is to spend the time
loading the target builds into their correct contexts, testing that, and
then finally running a test end-to-end of the entire pipeline.

There's a couple of TODO's here and there that need to be addressed as
well--but nothing too complex.